Refactor matrix utilities and operation tests
