+++ matrix_solvers/BUILD.bazel
@@ -7,11 +7,21 @@ cc_library(
     visibility = ["//visibility:public"],
 )
 
+cc_library(
+    name = "utilities_tests_lib",
+    hdrs = ["utilities_tests.h"],
+    deps = [
+        ":utilities",
+        "@googletest//:gtest_main",
+    ],
+)
+
 cc_test(
     name = "utilities_tests",
     srcs = ["utilities_tests.cpp"],
     deps = [
         ":utilities",
+        ":utilities_tests_lib",
         "@googletest//:gtest_main",
     ],
 )
@@ -29,6 +39,7 @@ cc_test(
     srcs = ["operations/operations_tests.cpp"],
     deps = [
         ":operations",
+        ":utilities_tests",
         "@googletest//:gtest_main",
     ],
 )
